requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.7
sqlalchemy>=2.0.23
setuptools<75
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
websockets==12.0
psycopg2-binary>=2.9.9
